\name{AFLPsim-package}
\alias{AFLPsim-package}
\encoding{latin1}
\docType{package}
\title{
Hybrid simulation and genome scan for dominant markers}

\description{
This package is developed in the Plant Reproductive Biology Lab (RNM-214) - University of Seville. It contains hybrid simulation functions for dominant genetic data. It also provides several genome scan methods. 

More information on AFLPsim can be found at \url{https://github.com/fbalao/AFLPsim}.

To cite AFLPsim, please use citation("AFLPsim")
}
\details{
\tabular{ll}{
Package: \tab AFLPsim\cr
Type: \tab Package\cr
Version: \tab 0.3-3\cr
Date: \tab 2013-03-22\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Francisco Balao; Juan Luis Garcia-Castano
}
\keyword{package}
