\name{AFLPsim-package}
\alias{AFLPsim-package}
\encoding{latin1}
\docType{package}
\title{
Hybrid Simulation and Genome Scan for Dominant Markers}

\description{
This package contains hybrid simulation functions for dominant genetic data. It also provides several genome scan methods. 

More information on AFLPsim can be found at \url{https://github.com/fbalao/AFLPsim}.

To cite AFLPsim, please use citation("AFLPsim")
}
\details{
\tabular{ll}{
Package: \tab AFLPsim\cr
Type: \tab Package\cr
Version: \tab 0.4-1\cr
Date: \tab 2013-12-24\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Francisco Balao; Juan Luis García-Castaño
}
\keyword{package}
